refactor: backend tests folders rearrangments
